% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/boxplotMeanDescriptor.mqic.R
\name{boxplotMeanDescriptor}
\alias{boxplotMeanDescriptor}
\title{boxplotMeanDescriptor: Draws box plots for a mean descriptor}
\usage{
boxplotMeanDescriptor(clsObject, meanDescriptor)
}
\arguments{
\item{clsObject}{nstance of mqic class}

\item{meanDescriptor}{Column name for the mean descriptor}
}
\description{
boxplotMeanDescriptor Function
}
\details{
Make a box plot (showing means, deviations) of a mean descriptor
(A1C, weight, BMI, FBG, SBP, DBP etc.) for patients with hypertension
or diabetes.
}
\examples{
\dontrun{
mqic <- mqic()
mqic <- readcsv(mqic, "MQIC.csv")
boxplotMeanDescriptor(mqic, "A1C_MEAN")
}
}

