% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/corrplotMeanDesciptor.mqic.R
\name{corrplotMeanDescriptor}
\alias{corrplotMeanDescriptor}
\title{corrplotMeanDescriptor: Draws a covariance matrix}
\usage{
corrplotMeanDescriptor(clsObject)
}
\arguments{
\item{clsObject}{Instance of mqic class}
}
\description{
corrplotMeanDescriptor Function
}
\details{
Displays a covariance matrix indicating the
correlations among the mean descriptors
}
\examples{
\dontrun{
mqic <- mqic()
readcsv(mqic, "MQIC.csv")
corrplotMeanDescriptor(mqic)
}
}

