% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/meanBMI.mqic.R
\name{meanBMI}
\alias{meanBMI}
\title{meanBMI: Returns the higest and the lowest BMI}
\usage{
meanBMI(clsObject)
}
\arguments{
\item{clsObject}{Instance of mqic class}
}
\description{
meanBMI Function
}
\details{
This function returns the highest and the lowest
mean BMI with the corresponding states.
}
\examples{
\dontrun{
mqic <- mqic()
mqic <- readcsv(mqic, "MQIC.csv")
meanBMI(mqic)
}
}

