% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/mqic.R
\name{mqic}
\alias{mqic}
\title{mqic: Defines mqic class}
\usage{
mqic()
}
\description{
mqic Class
}
\details{
This function defines mqic class to analyze MQIC
patient data.
}
\examples{
\dontrun{
mqic <- mqic()
}
}

