% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/patientCountByStates.mqic.R
\name{patientCountByStates}
\alias{patientCountByStates}
\title{patientCountByStates: Number of patients by states}
\usage{
patientCountByStates(clsObject, ...)
}
\arguments{
\item{clsObject}{Instance of mqic class}

\item{list}{disease, colNames, top}
}
\value{
countByStates Patient count by states
}
\description{
patientCountByStates Function
}
\details{
This function returns the number of patients by states.
}
\examples{
\dontrun{
mqic <- mqic()
mqic <- readcsv(mqic, "MQIC.csv")
patientCountByStates(mqic, list("disease"=1, "colNames"=c("State","TotalPatients"), "top"=5))
}
}

