% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/pieChartPatientCountByStates.mqic.R
\name{pieChartPatientCountByStates}
\alias{pieChartPatientCountByStates}
\title{pieChartPatientCountByStates: Draws a pie chart for the number of patients}
\usage{
pieChartPatientCountByStates(clsObject, disease = NULL)
}
\arguments{
\item{clsObject}{Instance of mqic class}

\item{disease}{Disease name}
}
\description{
pieChartPatientCountByStates Function
}
\details{
This function draws a pie chart for the
number of patients by states.
}
\examples{
\dontrun{
mqic <- mqic()
mqic <- readcsv(mqic, "MQIC.csv")
pieChartPatientCountByStates(mqic, "diabetes")
}
}

