% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/readcsv.mqic.R
\name{readcsv}
\alias{readcsv}
\title{readcsv: Draws a covariance matrix}
\usage{
readcsv(clsObject, data.file, filename = NULL)
}
\arguments{
\item{clsObject}{mqic Class instance.}

\item{data.file}{The name of the data file to be read.}
}
\value{
clsObject mqic Class instance.
}
\description{
readcsv Function
}
\details{
This function reads the comma separated values (.csv) file.
The data is loaded into the mqic instance as mqic$data.
}
\examples{
\dontrun{
mqic <- readcsv(mqic, "MQIC.csv")
}
}

