% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/scatterPlotCorrelate.mqic.R
\name{scatterPlotCorrelate}
\alias{scatterPlotCorrelate}
\title{scatterPlotCorrelate: Displays a scatterplot}
\usage{
scatterPlotCorrelate(clsObject, colx, coly)
}
\arguments{
\item{clsObject}{Instance of mqic class}

\item{colx}{Descriptor along the x axis}

\item{coly}{Descriptor along the y axis}
}
\description{
scatterPlotCorrelate Function
}
\details{
This function displays a scatterplot given
two descriptors.
}
\examples{
\dontrun{
mqic <- mqic()
mqic <- readcsv(mqic, "MQIC.csv")
scatterPlotCorrelate(mqic, "BMI_MEAN", "A1C_MEAN")
}
}

